# i18n: speed up replacement regex