The regex we use for checking if there are strings in the LHR that need to be replaced with their localized version can be really slow against giant strings. Adding a pre-pass with a fast regex is one easy approach to speeding this up without changing our ID format or being too particular about the file paths we allow.

In a trace from the issues we're seeing in #6067, time for localization and report generation goes from ~3500ms to ~33ms.

Even time for localization in the much smaller artifacts in `yarn update:sample-json` goes from 150ms to 22ms :)